Copy inlined slices gem from house seven
The existing code on the rails-4-i18n branch contains a bug which caused
data corruption on house seven. As we are updating other public site apps to
use Rails 4, we need the changes added to house seven ported to all the apps.

So it makes sense to un-inline the gem, as the initial argument was that not
much further dev work should be happening on any of the slices based apps. But
this has turned out to not be the case.